--HG--
branch : feature-streamed-package

+++ code/nel/src/misc/config_file/cf_lexical.cpp
@@ -2773,7 +2773,7 @@ static int input()
 
 		else
 			{ /* need more input */
-			int offset = yy_c_buf_p - yytext_ptr;
+			int offset = (int)(yy_c_buf_p - yytext_ptr);
 			++yy_c_buf_p;
 
 			switch ( yy_get_next_buffer() )
